pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-openai>=0.3.35",
    "langchain-text-splitters>=0.3.11",
    "pypdf>=6.1.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]
